# electro.client.react/src/libs/api-contract/swagger.json
{
	"openapi": "3.0.1",
	"info": {
	  "title": "WebAPI",
	  "version": "1.0"
	},
	"paths": {
...
	  "/api/opinions/products/{productId}": {
		"get": {
		  "tags": [
			"Opinions"
		  ],
		  "parameters": [
			{
			  "name": "productId",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			},
			{
			  "name": "ProductId",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			},
			{
			  "name": "Rating",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "Page",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "PageSize",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductOpinionsResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductOpinionsResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductOpinionsResult"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
...
	  "/api/products/{id}": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "id",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductResult"
				  }
				}
			  }
			},
			"404": {
			  "description": "Not Found",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products/{id}/similar": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "id",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			},
			{
			  "name": "limit",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32",
				"default": 8
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetSimilarProductsResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSimilarProductsResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSimilarProductsResult"
				  }
				}
			  }
			},
			"404": {
			  "description": "Not Found",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products/catalog": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "PageSize",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "Page",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductCatalogResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductCatalogResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetProductCatalogResult"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products": {
		"post": {
		  "tags": [
			"Products"
		  ],
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/CreateOrUpdateProductCommand"
				}
			  },
			  "text/json": {
				"schema": {
				  "$ref": "#/components/schemas/CreateOrUpdateProductCommand"
				}
			  },
			  "application/*+json": {
				"schema": {
				  "$ref": "#/components/schemas/CreateOrUpdateProductCommand"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/CreateOrUpdateProductResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/CreateOrUpdateProductResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/CreateOrUpdateProductResult"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products/search": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "GroupId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "CategoryId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "SubCategoryId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "Filters",
			  "in": "query",
			  "schema": {
				"type": "object",
				"additionalProperties": {
				  "type": "array",
				  "items": {
					"type": "string"
				  }
				}
			  }
			},
			{
			  "name": "Page",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "PageSize",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "Filters",
			  "in": "query",
			  "schema": {
				"type": "object",
				"additionalProperties": {
				  "type": "array",
				  "items": {
					"type": "string"
				  }
				}
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchProductsResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchProductsResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchProductsResult"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products/filters": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "GroupId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "CategoryId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			},
			{
			  "name": "SubCategoryId",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchFiltersResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchFiltersResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetSearchFiltersResult"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/products/bestsellers": {
		"get": {
		  "tags": [
			"Products"
		  ],
		  "parameters": [
			{
			  "name": "limit",
			  "in": "query",
			  "schema": {
				"type": "integer",
				"format": "int32",
				"default": 10
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/GetBestsellerProductsResult"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetBestsellerProductsResult"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/GetBestsellerProductsResult"
				  }
				}
			  }
			},
			"400": {
			  "description": "Bad Request",
			  "content": {
				"text/plain": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				},
				"text/json": {
				  "schema": {
					"$ref": "#/components/schemas/ProblemDetails"
				  }
				}
			  }
			}
		  }
		}
	  },
...
	},
	"components": {
	  "schemas": {
...
		"AttributeType": {
		  "enum": [
			"Text",
			"List",
			"Boolean"
		  ],
		  "type": "string"
		},
...
		"CreateOrUpdateProductCommand": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid",
			  "nullable": true
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"description": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photos": {
			  "type": "array",
			  "items": {
				"type": "string"
			  },
			  "nullable": true
			},
			"active": {
			  "type": "boolean"
			},
			"stockQuantity": {
			  "type": "integer",
			  "format": "int32"
			},
			"status": {
			  "type": "string",
			  "nullable": true
			},
			"groupId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"categoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"subCategoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"attributes": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/CreateOrUpdateProductCommandAttributeDefinitionValue"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"CreateOrUpdateProductCommandAttributeDefinitionValue": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"value": {
			  "type": "string",
			  "nullable": true
			},
			"isPrimary": {
			  "type": "boolean"
			}
		  },
		  "additionalProperties": false
		},
		"CreateOrUpdateProductResult": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"description": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photos": {
			  "type": "array",
			  "items": {
				"type": "string"
			  },
			  "nullable": true
			},
			"status": {
			  "type": "string",
			  "nullable": true
			},
			"groupId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"categoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"subCategoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"attributes": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/ProductAttributeResult"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
...
		"GetBestsellerProductsResult": {
		  "type": "object",
		  "properties": {
			"products": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetBestsellerProductsResultProduct"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"GetBestsellerProductsResultProduct": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photo": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
...
		"GetProductCatalogResult": {
		  "type": "object",
		  "properties": {
			"products": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetProductCatalogResultProduct"
			  },
			  "nullable": true
			},
			"pageCount": {
			  "type": "integer",
			  "format": "int32"
			},
			"pageSize": {
			  "type": "integer",
			  "format": "int32"
			},
			"page": {
			  "type": "integer",
			  "format": "int32"
			}
		  },
		  "additionalProperties": false
		},
		"GetProductCatalogResultProduct": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"price": {
			  "$ref": "#/components/schemas/Money"
			},
			"photo": {
			  "type": "string",
			  "nullable": true
			},
			"status": {
			  "$ref": "#/components/schemas/ProductStatus"
			},
			"stockQuantity": {
			  "type": "integer",
			  "format": "int32"
			}
		  },
		  "additionalProperties": false
		},
		"GetProductOpinionsResult": {
		  "type": "object",
		  "properties": {
			"items": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetProductOpinionsResultOpinion"
			  },
			  "nullable": true
			},
			"page": {
			  "type": "integer",
			  "format": "int32"
			},
			"pageSize": {
			  "type": "integer",
			  "format": "int32"
			},
			"totalPages": {
			  "type": "integer",
			  "format": "int32",
			  "readOnly": true
			}
		  },
		  "additionalProperties": false
		},
		"GetProductOpinionsResultOpinion": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"content": {
			  "type": "string",
			  "nullable": true
			},
			"rating": {
			  "type": "number",
			  "format": "float"
			},
			"createdAt": {
			  "type": "string",
			  "format": "date-time"
			},
			"authorDisplayName": {
			  "type": "string",
			  "nullable": true
			},
			"reactionType": {
			  "$ref": "#/components/schemas/OpinionReactionType"
			},
			"likesCount": {
			  "type": "integer",
			  "format": "int32"
			},
			"dislikesCount": {
			  "type": "integer",
			  "format": "int32"
			}
		  },
		  "additionalProperties": false
		},
...
		"GetProductResult": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"description": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photos": {
			  "type": "array",
			  "items": {
				"type": "string"
			  },
			  "nullable": true
			},
			"status": {
			  "$ref": "#/components/schemas/ProductStatus"
			},
			"groupId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"categoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"subCategoryId": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"active": {
			  "type": "boolean"
			},
			"stockQuantity": {
			  "type": "integer",
			  "format": "int32"
			},
			"averageOpinionRating": {
			  "type": "number",
			  "format": "float"
			},
			"opinionCount": {
			  "type": "integer",
			  "format": "int32"
			},
			"attributes": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/ProductAttributeResult"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
...
		"GetSearchFiltersResult": {
		  "type": "object",
		  "properties": {
			"filters": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetSearchFiltersResultElement"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"GetSearchFiltersResultElement": {
		  "type": "object",
		  "properties": {
			"attributeDefinitionId": {
			  "type": "string",
			  "format": "uuid"
			},
			"type": {
			  "$ref": "#/components/schemas/AttributeType"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"values": {
			  "type": "array",
			  "items": {
				"type": "string"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"GetSearchProductsResult": {
		  "type": "object",
		  "properties": {
			"items": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetSearchProductsResultProduct"
			  },
			  "nullable": true
			},
			"page": {
			  "type": "integer",
			  "format": "int32"
			},
			"pageSize": {
			  "type": "integer",
			  "format": "int32"
			},
			"totalPages": {
			  "type": "integer",
			  "format": "int32",
			  "readOnly": true
			}
		  },
		  "additionalProperties": false
		},
		"GetSearchProductsResultProduct": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photo": {
			  "type": "string",
			  "nullable": true
			},
			"status": {
			  "$ref": "#/components/schemas/ProductStatus"
			},
			"averageOpinionRating": {
			  "type": "number",
			  "format": "float"
			},
			"opinionCount": {
			  "type": "integer",
			  "format": "int32"
			},
			"attributes": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/ProductAttributeResult"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"GetSimilarProductsResult": {
		  "type": "object",
		  "properties": {
			"products": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/GetSimilarProductsResultProduct"
			  },
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"GetSimilarProductsResultProduct": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			},
			"photo": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
...
		"Money": {
		  "type": "object",
		  "properties": {
			"amount": {
			  "type": "number",
			  "format": "double"
			},
			"currency": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"OpinionReactionType": {
		  "enum": [
			"Like",
			"Dislike"
		  ],
		  "type": "string"
		},
...
		"ProblemDetails": {
		  "type": "object",
		  "properties": {
			"type": {
			  "type": "string",
			  "nullable": true
			},
			"title": {
			  "type": "string",
			  "nullable": true
			},
			"status": {
			  "type": "integer",
			  "format": "int32",
			  "nullable": true
			},
			"detail": {
			  "type": "string",
			  "nullable": true
			},
			"instance": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": { }
		},
		"ProductAttributeResult": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"name": {
			  "type": "string",
			  "nullable": true
			},
			"type": {
			  "$ref": "#/components/schemas/AttributeType"
			},
			"value": {
			  "type": "string",
			  "nullable": true
			},
			"isPrimary": {
			  "type": "boolean"
			}
		  },
		  "additionalProperties": false
		},
		"ProductStatus": {
		  "enum": [
			"Draft",
			"Active",
			"Inactive",
			"Discontinued"
		  ],
		  "type": "string"
		},
...
	  },
	  "securitySchemes": {
		"Bearer": {
		  "type": "http",
		  "scheme": "Bearer"
		}
	  }
	},
	"security": [
	  {
		"Bearer": [ ]
	  }
	]
  }